b/f: core: use %zu instead where possible replaced %lu and c-style cast with %zu in cases of STL container size()

+++ core/AmB2BMedia.cpp
@@ -845,8 +845,7 @@ void AmB2BMedia::replaceConnectionAddress(AmSdp &parser_sdp, bool a_leg,
   if (it != parser_sdp.media.end()) {
     // FIXME: create new streams here?
     WARN("trying to relay SDP with more media lines than "
-	 "relay streams initialized (%lu)\n", 
-	 (unsigned long)(audio.size()+relay_streams.size()));
+	 "relay streams initialized (%zu)\n",audio.size()+relay_streams.size());
   }
 
   DBG("replaced connection address in SDP with %s:%s.\n",
